requirements.txt
aiogram~=2.15.0
python-dotenv
sqlalchemy[asyncio]
aiosqlite
asyncpg